# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "sms-app api documentation",
        "description": "sms-app api documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/contacts": {
            "get": {
                "tags": [
                    "Show All Contacts"
                ],
                "description": "Show all contacts",
                "operationId": "AllContacts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry no contact found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Create Contact"
                ],
                "description": "Create new contact",
                "operationId": "CreateContact",
                "parameters": [
                    {
                        "name": "sur_name",
                        "in": "query",
                        "description": "Enter sur name of the contact person",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Enter first name of the contact person",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter email address of the contact person",
                        "required": true
                    },
                    {
                        "name": "scheme",
                        "in": "query",
                        "description": "Enter scheme of the contact person",
                        "required": true
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Enter gender of the contact person",
                        "required": true
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "description": "Enter date of birth of the contact person",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact created successfully"
                    },
                    "400": {
                        "description": "Sorry unable to create contact"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/contacts-batch": {
            "post": {
                "tags": [
                    "UploadContact"
                ],
                "description": "Upload batch contacts",
                "operationId": "UploadBatchContacts",
                "parameters": [
                    {
                        "name": "file",
                        "in": "query",
                        "description": "Csv file is required",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacts uploaded successfully"
                    },
                    "400": {
                        "description": "Sorry unable to create contact"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "tags": [
                    "Show Contact"
                ],
                "description": "Show single contact",
                "operationId": "ShowContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Enter contact id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Sorry this data do not exist"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete Contact"
                ],
                "summary": "Delete existing contact",
                "description": "Deletes a contact and returns no content",
                "operationId": "deleteContact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Sorry this contact do not exist"
                    },
                    "400": {
                        "description": "Sorry unable to delete contact"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Update Contacts"
                ],
                "operationId": "UpdateContacts",
                "parameters": [
                    {
                        "name": "sur_name",
                        "in": "query",
                        "description": "Enter sur name of the contact person",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Enter first name of the contact person",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Enter email address of the contact person",
                        "required": true
                    },
                    {
                        "name": "scheme",
                        "in": "query",
                        "description": "Enter scheme of the contact person",
                        "required": true
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Enter gender of the contact person",
                        "required": true
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "description": "Enter date of birth of the contact person",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact updated successfully"
                    },
                    "400": {
                        "description": "Sorry unable to update contact"
                    },
                    "404": {
                        "description": "Sorry this data do not exist"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Contact": {
                "title": "Contact",
                "description": "Contact model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "sur_name": {
                        "title": "Sur name",
                        "description": "Sur name of the contact person ",
                        "type": "string",
                        "example": "Orutu"
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "Frist name of contact person",
                        "type": "string",
                        "example": "Akposieyefa"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the contact person",
                        "type": "string",
                        "example": "orutu1@gmail.com"
                    },
                    "scheme": {
                        "title": "Scheme",
                        "description": "Enter scheme of the contact person",
                        "type": "string",
                        "example": "NHIS"
                    },
                    "gender": {
                        "title": "Gender",
                        "description": "Enter gender of the contact person",
                        "type": "string",
                        "example": "Male"
                    },
                    "dob": {
                        "title": "DOB",
                        "description": "Date of birth of the contact person",
                        "type": "string",
                        "example": "1995-06-15"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contact"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}